move BatchedGraph back to model.py
