+++ chgnet/graph/__init__.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
 from chgnet.graph.converter import CrystalGraphConverter
-from chgnet.graph.crystalgraph import BatchedGraph, CrystalGraph
+from chgnet.graph.crystalgraph import CrystalGraph
 
-__all__ = ["CrystalGraph", "BatchedGraph", "CrystalGraphConverter"]
+__all__ = ["CrystalGraph", "CrystalGraphConverter"]
