+++ chgnet/graph/crystalgraph.py
@@ -1,11 +1,10 @@
 from __future__ import annotations
 
 import os
-from dataclasses import dataclass
-from typing import Any, Sequence
+from typing import Any
 
 import torch
-from torch import Tensor, nn
+from torch import Tensor
 
 datatype = torch.float32
 
@@ -145,161 +144,3 @@ class CrystalGraph:
 
     def __repr__(self) -> str:
         return str(self)
-
-
-@dataclass
-class BatchedGraph:
-    """Batched crystal graph for parallel computing.
-
-    Attributes:
-        atomic_numbers (Tensor): atomic numbers vector
-            [num_batch_atoms]
-        bond_bases_ag (Tensor): bond bases vector for atom_graph
-            [num_batch_bonds, num_radial]
-        bond_bases_bg (Tensor): bond bases vector for atom_graph
-            [num_batch_bonds, num_radial]
-        angle_bases (Tensor): angle bases vector
-            [num_batch_angles, num_angular]
-        batched_atom_graph (Tensor) : batched atom graph adjacency list
-            [num_batch_bonds, 2]
-        batched_bond_graph (Tensor) : bond graph adjacency list
-            [num_batch_angles, 2]
-        atom_owners (Tensor): graph indices for each atom, used aggregate batched
-            graph back to single graph
-            [num_batch_atoms]
-        directed2undirected (Tensor): the utility tensor used to quickly
-            map directed edges to undirected edges in graph
-            [num_directed]
-        atom_positions (List[Tensor]): cartesian coordinates of the atoms
-            from structures
-            [num_batch_atoms]
-        strains (List[Tensor]): a list of strains that's initialized to be zeros
-            [batch_size]
-        volumes (Tensor): the volume of each structure in the batch
-            [batch_size]
-    """
-
-    atomic_numbers: Tensor
-    bond_bases_ag: Tensor
-    bond_bases_bg: Tensor
-    angle_bases: Tensor
-    batched_atom_graph: Tensor
-    batched_bond_graph: Tensor
-    atom_owners: Tensor
-    directed2undirected: Tensor
-    atom_positions: Sequence[Tensor]
-    strains: Sequence[Tensor]
-    volumes: Sequence[Tensor]
-
-    @classmethod
-    def from_graphs(
-        cls,
-        graphs: Sequence[CrystalGraph],
-        bond_basis_expansion: nn.Module,
-        angle_basis_expansion: nn.Module,
-        compute_stress: bool = False,
-    ) -> BatchedGraph:
-        """Featurize and assemble a list of graphs.
-
-        Args:
-            graphs (List[Tensor]): a list of Crystal_Graphs
-            bond_basis_expansion (nn.Module): bond basis expansion layer in CHGNet
-            angle_basis_expansion (nn.Module): angle basis expansion layer in CHGNet
-            compute_stress (bool): whether to compute stress
-
-        Returns:
-            assembled batch_graph that is ready for batched forward pass in CHGNet
-        """
-        atomic_numbers, atom_positions = [], []
-        strains, volumes = [], []
-        bond_bases_ag, bond_bases_bg, angle_bases = [], [], []
-        batched_atom_graph, batched_bond_graph = [], []
-        directed2undirected = []
-        atom_owners = []
-        atom_offset_idx = 0
-        n_undirected = 0
-
-        for graph_idx, graph in enumerate(graphs):
-            # Atoms
-            n_atom = graph.atomic_number.shape[0]
-            atomic_numbers.append(graph.atomic_number)
-
-            # Lattice
-            if compute_stress:
-                strain = graph.lattice.new_zeros([3, 3], requires_grad=True)
-                lattice = graph.lattice @ (torch.eye(3).to(strain.device) + strain)
-            else:
-                strain = None
-                lattice = graph.lattice
-            volumes.append(torch.det(lattice))
-            strains.append(strain)
-
-            # Bonds
-            atom_cart_coords = graph.atom_frac_coord @ lattice
-            bond_basis_ag, bond_basis_bg, bond_vectors = bond_basis_expansion(
-                center=atom_cart_coords[graph.atom_graph[:, 0]],
-                neighbor=atom_cart_coords[graph.atom_graph[:, 1]],
-                undirected2directed=graph.undirected2directed,
-                image=graph.neighbor_image,
-                lattice=lattice,
-            )
-            atom_positions.append(atom_cart_coords)
-            bond_bases_ag.append(bond_basis_ag)
-            bond_bases_bg.append(bond_basis_bg)
-
-            # Indexes
-            batched_atom_graph.append(graph.atom_graph + atom_offset_idx)
-            directed2undirected.append(graph.directed2undirected + n_undirected)
-
-            # Angles
-            if len(graph.bond_graph) != 0:
-                bond_vecs_i = torch.index_select(
-                    bond_vectors, 0, graph.bond_graph[:, 2]
-                )
-                bond_vecs_j = torch.index_select(
-                    bond_vectors, 0, graph.bond_graph[:, 4]
-                )
-                angle_basis = angle_basis_expansion(bond_vecs_i, bond_vecs_j)
-                angle_bases.append(angle_basis)
-
-                bond_graph = graph.bond_graph.new_zeros([graph.bond_graph.shape[0], 3])
-                bond_graph[:, 0] = graph.bond_graph[:, 0] + atom_offset_idx
-                bond_graph[:, 1] = graph.bond_graph[:, 1] + n_undirected
-                bond_graph[:, 2] = graph.bond_graph[:, 3] + n_undirected
-                batched_bond_graph.append(bond_graph)
-
-            atom_owners.append(torch.ones(n_atom, requires_grad=False) * graph_idx)
-            atom_offset_idx += n_atom
-            n_undirected += len(bond_basis_ag)
-
-        # Make Torch Tensors
-        atomic_numbers = torch.cat(atomic_numbers, dim=0)
-        bond_bases_ag = torch.cat(bond_bases_ag, dim=0)
-        bond_bases_bg = torch.cat(bond_bases_bg, dim=0)
-        angle_bases = (
-            torch.cat(angle_bases, dim=0) if len(angle_bases) != 0 else torch.tensor([])
-        )
-        batched_atom_graph = torch.cat(batched_atom_graph, dim=0)
-        if batched_bond_graph != []:
-            batched_bond_graph = torch.cat(batched_bond_graph, dim=0)
-        else:  # when bond graph is empty or disabled
-            batched_bond_graph = torch.tensor([])
-        atom_owners = (
-            torch.cat(atom_owners, dim=0).type(torch.int).to(atomic_numbers.device)
-        )
-        directed2undirected = torch.cat(directed2undirected, dim=0)
-        volumes = torch.tensor(volumes, dtype=datatype, device=atomic_numbers.device)
-
-        return cls(
-            atomic_numbers=atomic_numbers,
-            bond_bases_ag=bond_bases_ag,
-            bond_bases_bg=bond_bases_bg,
-            angle_bases=angle_bases,
-            batched_atom_graph=batched_atom_graph,
-            batched_bond_graph=batched_bond_graph,
-            atom_owners=atom_owners,
-            directed2undirected=directed2undirected,
-            atom_positions=atom_positions,
-            strains=strains,
-            volumes=volumes,
-        )
